requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-translate==3.15.1
google-cloud-pubsub==2.18.4
google-cloud-firestore==2.13.1
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0

